+++ src/Control/Concurrent/NQE/Process.hs
@@ -3,162 +3,252 @@
 {-# LANGUAGE FlexibleInstances         #-}
 {-# LANGUAGE MultiParamTypeClasses     #-}
 {-# LANGUAGE RankNTypes                #-}
-module Control.Concurrent.NQE.Process
-    ( Mailbox(..)
-    , Inbox
-    , Reply
-    , Listen
-    , newInbox
-    , send
-    , receive
-    , query
-    , receiveMatch
-    , receiveMatchSTM
-    , mailboxEmpty
-    ) where
+module Control.Concurrent.NQE.Process where
 
 import           Control.Concurrent.Unique
+import           Control.Monad
+import           Data.Dynamic
+import           Data.Function
 import           Data.Hashable
 import           UnliftIO
 
--- | STM action to reply to a synchronous request.
+-- | Wrapper for STM action to send a reply to a request.
 type Reply a = a -> STM ()
 
--- | STM action for an event listener.
-type Listen a = a -> STM ()
+-- | Mailbox view that only allows to send messages.
+data Mailbox =
+    forall mbox. (OutChan mbox) =>
+                 Mailbox !mbox
+                         !Unique
 
--- | Mailboxes are used to communicate with processes (actors). A process will
--- usually listen in a loop for events entering its mailbox. A process is its
--- mailbox, so it may be named as the process that it communicates with.
---
--- >>> :m + Control.Monad NQE UnliftIO
--- >>> registry <- newTQueueIO :: IO (TQueue String)
--- >>> let run = receive registry >>= putStrLn . ("Registered: " ++)
--- >>> withAsync run $ \a -> "Bruce Wayne" `send` registry >> wait a
--- Registered: Bruce Wayne
-class Eq (mbox msg) => Mailbox mbox msg where
-    -- | STM action that responds true if the mailbox is empty. Useful to avoid
-    -- blocking on an empty mailbox.
-    mailboxEmptySTM :: mbox msg -> STM Bool
-    -- | STM action that responds true if the mailbox is full and would block if
-    -- a new message is received.
-    mailboxFullSTM :: mbox msg -> STM Bool
-    -- | STM action to send a message to a mailbox. This is usually called from
-    -- a process that wishes to communicate with actor that owns the mailbox.
-    sendSTM :: msg -> mbox msg -> STM ()
-    -- | STM action to receive a message from a mailbox. This should be called
-    -- from the process that owns the mailbox.
-    receiveSTM :: mbox msg -> STM msg
-    -- | Put a message back in the mailbox so that it is the next one to be
-    -- received. Used for pattern matching.
-    requeueSTM :: msg -> mbox msg -> STM ()
+-- | Function for sorting dispatchers for different types of incoming message.
+data Dispatch m =
+    forall msg. (Typeable msg) => Dispatch { getDispatcher :: msg -> m () }
 
-instance Mailbox TQueue msg where
+-- | Mailbox for a process allowing sending and receiving messages.
+data Inbox =
+    forall mbox. (OutChan mbox, InChan mbox) =>
+                 Inbox !mbox
+                       !Unique
+
+instance Eq Mailbox where
+    (==) = (==) `on` f
+      where
+        f (Mailbox _ u) = u
+
+instance Eq Inbox where
+    (==) = (==) `on` f
+      where
+        f (Inbox _ u) = u
+
+-- | Thread asynchronous handle and its mailbox.
+data Process = Process
+    { getProcessAsync   :: Async ()
+    , getProcessMailbox :: Mailbox
+    } deriving (Typeable)
+
+instance Eq Process where
+    (==) = (==) `on` getProcessMailbox
+
+-- | Class for channels that can be used to build mailboxes.
+class (Eq mbox, Typeable mbox) => InChan mbox where
+    -- | Are there pending messages?
+    mailboxEmptySTM :: mbox -> STM Bool
+    -- | Receive a 'Dynamic' from mailbox in an 'STM' transaction.
+    receiveDynSTM :: mbox -> STM Dynamic
+    -- | Re-enqueue a 'Dynamic' in the mailbox such that it is the next element
+    -- to be retrieved by 'receiveDynSTM'.
+    requeueDynSTM :: Dynamic -> mbox -> STM ()
+
+class (Eq mbox, Typeable mbox) => OutChan mbox where
+    -- | Is this bounded mailbox full?
+    mailboxFullSTM :: mbox -> STM Bool
+    -- | Send a 'Dynamic' to the mailbox in an 'STM' transaction.
+    sendDynSTM :: Dynamic -> mbox -> STM ()
+
+instance InChan (TQueue Dynamic) where
     mailboxEmptySTM = isEmptyTQueue
-    mailboxFullSTM = const $ return False
-    sendSTM msg = (`writeTQueue` msg)
-    receiveSTM = readTQueue
-    requeueSTM msg = (`unGetTQueue` msg)
+    receiveDynSTM = readTQueue
+    requeueDynSTM msg = (`unGetTQueue` msg)
 
-instance Mailbox TBQueue msg where
+instance OutChan (TQueue Dynamic) where
+    mailboxFullSTM _ = return False
+    sendDynSTM msg = (`writeTQueue` msg)
+
+instance InChan (TBQueue Dynamic) where
     mailboxEmptySTM = isEmptyTBQueue
+    receiveDynSTM = readTBQueue
+    requeueDynSTM msg = (`unGetTBQueue` msg)
+
+instance OutChan (TBQueue Dynamic) where
     mailboxFullSTM = isFullTBQueue
-    sendSTM msg = (`writeTBQueue` msg)
-    receiveSTM = readTBQueue
-    requeueSTM msg = (`unGetTBQueue` msg)
+    sendDynSTM msg = (`writeTBQueue` msg)
 
--- | Wrapped 'Mailbox' hiding its implementation.
-data Inbox msg = forall mbox. (Mailbox mbox msg) => Inbox !(mbox msg) !Unique
+instance OutChan Mailbox where
+    mailboxFullSTM (Mailbox mbox _) = mailboxFullSTM mbox
+    sendDynSTM msg (Mailbox mbox _) = msg `sendDynSTM` mbox
 
--- | Create a new 'Inbox' with a 'Unique' identifier inside. If you run this
--- function more than once with the same 'Mailbox', its results will be
--- different from the 'Eq' or 'Hashable' point of view.
-newInbox :: (MonadIO m, Mailbox mbox msg) => mbox msg -> m (Inbox msg)
+instance InChan Inbox where
+    mailboxEmptySTM (Inbox mbox _) = mailboxEmptySTM mbox
+    receiveDynSTM (Inbox mbox _) = receiveDynSTM mbox
+    requeueDynSTM msg (Inbox mbox _) = msg `requeueDynSTM` mbox
+
+instance OutChan Inbox where
+    mailboxFullSTM (Inbox mbox _) = mailboxFullSTM mbox
+    sendDynSTM msg (Inbox mbox _) = msg `sendDynSTM` mbox
+
+instance OutChan Process where
+    mailboxFullSTM (Process _ mbox) = mailboxFullSTM mbox
+    sendDynSTM msg (Process _ mbox) = msg `sendDynSTM` mbox
+
+instance Hashable Process where
+    hashWithSalt i (Process _ m) = hashWithSalt i m
+    hash (Process _ m) = hash m
+
+instance Hashable Mailbox where
+    hashWithSalt i (Mailbox _ u) = hashWithSalt i u
+    hash (Mailbox _ u) = hash u
+
+-- | Remove read capabilities from an 'Inbox' to get a write-only 'Mailbox'.
+inboxToMailbox :: Inbox -> Mailbox
+inboxToMailbox (Inbox m u) = Mailbox m u
+
+-- | Wrap a bi-directional channel in an 'Inbox'.
+newInbox :: (MonadIO m, InChan mbox, OutChan mbox) => mbox -> m Inbox
 newInbox mbox = Inbox mbox <$> liftIO newUnique
 
-instance Eq (Inbox msg) where
-    Inbox _ u1 == Inbox _ u2 = u1 == u2
-
-instance Hashable (Inbox msg) where
-    hashWithSalt i (Inbox _ u) = hashWithSalt i u
-    hash (Inbox _ u) = hash u
-
-instance Mailbox Inbox msg where
-    mailboxEmptySTM (Inbox mbox _) = mailboxEmptySTM mbox
-    mailboxFullSTM (Inbox mbox _) = mailboxFullSTM mbox
-    sendSTM msg (Inbox mbox _) = msg `sendSTM` mbox
-    receiveSTM (Inbox mbox _) = receiveSTM mbox
-    requeueSTM msg (Inbox mbox _) = msg `requeueSTM` mbox
-
--- | Send a message to a mailbox.
-send :: (MonadIO m, Mailbox mbox msg) => msg -> mbox msg -> m ()
+-- | Send a message to a 'Mailbox'.
+send :: (MonadIO m, Typeable msg, OutChan mbox) => msg -> mbox -> m ()
 send msg = atomically . sendSTM msg
 
--- | Receive a message from the mailbox. This function should be called only by
--- the process that owns the malibox.
-receive ::
-       (MonadIO m, Mailbox mbox msg)
-    => mbox msg
-    -> m msg
+-- | Send a message to a 'Mailbox' in an 'STM' transaction.
+sendSTM :: (Typeable msg, OutChan mbox) => msg -> mbox -> STM ()
+sendSTM msg = sendDynSTM (toDyn msg)
+
+-- | Receive a message from a 'Mailbox'. Will block until a message of the right
+-- type appears. Any message of the wrong type will remain in the mailbox.
+receive :: (Typeable msg, InChan mbox, MonadIO m) => mbox -> m msg
 receive mbox = receiveMatch mbox Just
 
--- | Use a partially-applied message type that takes a `Reply a` as its last
--- argument. This function will create the STM action for the response, send the
--- message to a process and await for the STM action to be fulfilled before
--- responding response. It implements synchronous communication with a process.
---
--- Example:
---
--- >>> :m + NQE UnliftIO
--- >>> data Message = Square Integer (Reply Integer)
--- >>> doubler <- newTQueueIO :: IO (TQueue Message)
--- >>> let proc = receive doubler >>= \(Square i r) -> atomically $ r (i * i)
--- >>> withAsync proc $ \_ -> Square 2 `query` doubler
--- 4
---
--- In this example the @Square@ constructor takes a 'Reply' action as its last
--- argument. It is passed partially-applied to @query@, which adds a new 'Reply'
--- action before sending it to the @doubler@ and then waiting for it. The
--- doubler process will run the @Reply@ action n STM with the reply as its
--- argument. In this case @i * i@.
+-- | Receive a message from a 'Mailbox' in an 'STM' transaction. Will block
+-- until a message of the right type appears. Any message of the wrong type will
+-- remain in the mailbox.
+receiveSTM :: (Typeable msg, InChan mbox) => mbox -> STM msg
+receiveSTM mbox = receiveMatchSTM mbox Just
+
+receiveDyn :: (MonadIO m, InChan mbox) => mbox -> m Dynamic
+receiveDyn = atomically . receiveDynSTM
+
+-- | Dispatch incoming message according to its type.
+-- | If message cannot be dispatched, discard or run default action if present.
+dispatch ::
+       (MonadIO m, InChan mbox)
+    => Maybe (Dynamic -> m ()) -- ^ action if cannot dispatch
+    -> mbox
+    -> [Dispatch m]
+    -> m ()
+dispatch md mbox ds = join $ atomically (dispatchSTM md mbox ds)
+
+dispatchSTM ::
+       (Applicative m, InChan mbox)
+    => Maybe (Dynamic -> m ())
+    -> mbox
+    -> [Dispatch m]
+    -> STM (m ())
+dispatchSTM md mbox ds = receiveDynSTM mbox >>= go ds
+  where
+    go [] msg = return $ maybe (pure ()) ($ msg) md
+    go (Dispatch f:xs) msg =
+        case fromDynamic msg of
+            Just a  -> return (f a)
+            Nothing -> go xs msg
+
+-- | Send request to mailbox and wait for a response. Provide a function that
+-- takes an empty 'TMVar' and produces a request that is sent to the mailbox.
+-- The remote process should fulfill the 'TMVar'. At this point this function
+-- will return its contents.
 query ::
-       (MonadIO m, Mailbox mbox msg)
-    => (Reply a -> msg)
-    -> mbox msg
-    -> m a
-query f mbox = do
-    box <- atomically newEmptyTMVar
-    f (putTMVar box) `send` mbox
-    atomically (takeTMVar box)
+       (MonadIO m, Typeable request, OutChan mbox)
+    => (Reply response -> request)
+    -> mbox
+    -> m response
+query f m = do
+    r <- newEmptyTMVarIO
+    f (putTMVar r) `send` m
+    atomically $ takeTMVar r
+
+-- | Do a 'query' but timeout after @u@ microseconds. Return 'Nothing' if
+-- timeout reached.
+queryU ::
+       (MonadUnliftIO m, Typeable request, OutChan mbox)
+    => Int
+    -> (Reply response -> request)
+    -> mbox
+    -> m (Maybe response)
+queryU u f m = timeout u (query f m)
+
+-- | Do a 'query' but timeout after @s@ seconds. Return 'Nothing' if
+-- timeout reached.
+queryS ::
+       (MonadUnliftIO m, Typeable request, OutChan mbox)
+    => Int
+    -> (Reply response -> request)
+    -> mbox
+    -> m (Maybe response)
+queryS s f m = timeout s (query f m)
+
+-- | Do a 'query' but timeout after 30 seconds. Return 'Nothing' if
+-- timeout reached.
+query30 ::
+       (MonadUnliftIO m, Typeable request, OutChan mbox)
+    => (Reply response -> request)
+    -> mbox
+    -> m (Maybe response)
+query30 = queryS 30
 
 -- | Test all the messages in a mailbox against the supplied function and return
--- the output of the function only when it is 'Just'. Will block until a message
--- matches. All messages that did not match are left in the mailbox. Only call
--- from process that owns mailbox.
+-- the matching message. Will block until a match is found. Messages that do not
+-- match remain in the mailbox.
 receiveMatch ::
-       (MonadIO m, Mailbox mbox msg) => mbox msg -> (msg -> Maybe a) -> m a
+       (MonadIO m, Typeable msg, InChan mbox) => mbox -> (msg -> Maybe a) -> m a
 receiveMatch mbox = atomically . receiveMatchSTM mbox
 
 -- | Match a message in the mailbox as an atomic STM action.
 receiveMatchSTM ::
-       (Mailbox mbox msg)
-    => mbox msg
-    -> (msg -> Maybe a)
-    -> STM a
+       (Typeable msg, InChan mbox) => mbox -> (msg -> Maybe a) -> STM a
 receiveMatchSTM mbox f = go []
   where
     go acc =
-        receiveSTM mbox >>= \msg ->
-            case f msg of
+        receiveDynSTM mbox >>= \msg ->
+            case fromDynamic msg >>= f of
                 Just x -> do
                     requeueListSTM acc mbox
                     return x
                 Nothing -> go (msg : acc)
 
 -- | Check if the mailbox is empty.
-mailboxEmpty :: (MonadIO m, Mailbox mbox msg) => mbox msg -> m Bool
+mailboxEmpty :: (MonadIO m, InChan mbox) => mbox -> m Bool
 mailboxEmpty = atomically . mailboxEmptySTM
 
 -- | Put a message at the start of a mailbox, so that it is the next one read.
-requeueListSTM :: (Mailbox mbox msg) => [msg] -> mbox msg -> STM ()
-requeueListSTM xs mbox = mapM_ (`requeueSTM` mbox) xs
+requeueListSTM :: InChan mbox => [Dynamic] -> mbox -> STM ()
+requeueListSTM xs mbox = mapM_ (`requeueDynSTM` mbox) xs
+
+-- | Run a process in the background.
+withProcess :: MonadUnliftIO m => (Inbox -> m ()) -> (Process -> m a) -> m a
+withProcess p f = do
+    (i, m) <- newMailbox
+    withAsync (p i) (\a -> f (Process a m))
+
+process :: MonadUnliftIO m => (Inbox -> m ()) -> m Process
+process p = do
+    (i, m) <- newMailbox
+    a <- async $ p i
+    return (Process a m)
+
+newMailbox :: MonadUnliftIO m => m (Inbox, Mailbox)
+newMailbox = do
+    q <- newTQueueIO
+    i <- newInbox (q :: TQueue Dynamic)
+    let m = inboxToMailbox i
+    return (i, m)